backend-ai/requirements.txt
flask
flask-cors
google-generativeai
python-dotenv
Pillow==10.1.0
